subir

+++ src/main.ts
@@ -8,7 +8,7 @@ import { filtrarTodas } from "./funcionalidades-Puras/FiltroSegunEstado";
 import { pedirNumero } from "./Funcionalidades/Verificadores";
 import { BusquedaTitulo } from "./funcionalidades-Puras/FiltroSegunEstado";
 import { mostarTitulos, mostrarTareaCompletas } from "./Interfaz/ConsolaTarea";
-import { manejoEstadisticas } from "./logica/FlujodeNegocio";
+import { manejoConsultas, manejoEstadisticas } from "./logica/FlujodeNegocio";
 function main(): void {
 
 
@@ -41,18 +41,17 @@ function main(): void {
                     console.log(" No hay tareas disponibles.");
                 }
                 else {
-                console.log(mostarTitulos(gestorTareas.getItems() , gestorTareas.getItems().length , 0));
-                let tituloBuscar: string
-                tituloBuscar= prompt("\n Ingrese el título de la tarea a buscar: ").trim();
-                const tareaEncontrada = BusquedaTitulo(gestorTareas.getItems(), tituloBuscar);
-                if (tareaEncontrada) {
-                    console.log("Tarea encontrada:");
-                    console.log(mostrarTareaCompletas(tareaEncontrada));
-                } else {
-                    console.log("No se encontró ninguna tarea con ese título.");
-                }
-                pausar();
-                break;
+                    console.log(mostarTitulos(gestorTareas.getItems() , gestorTareas.getItems().length , 0));
+                    let tituloBuscar: string
+                    tituloBuscar= prompt("\n Ingrese el título de la tarea a buscar: ").trim();
+                    const tareaEncontrada = BusquedaTitulo(gestorTareas.getItems(), tituloBuscar);
+                    if (tareaEncontrada) {
+                        console.log("Tarea encontrada:");
+                        console.log(mostrarTareaCompletas(tareaEncontrada));
+                    } else {
+                        console.log("No se encontró ninguna tarea con ese título.");
+                    }
+                    pausar();
                 }
             break;
 
@@ -65,39 +64,52 @@ function main(): void {
             break;
             
             case 4:
+                if(gestorTareas.getItems().length == 0){
+                    console.log(" No hay tareas disponibles.");
+                    pausar();
+                }else{
+                    let tareas=filtrarTodas(gestorTareas.getItems());
+                    tareas.forEach(tarea => {
+                        console.log(tarea.toString());
+                        console.log("----------------------------------------");
+                    });
+                    console.log(`${tareas.length+1}. Salir del eliminar`)
 
-                console.log("\n=== ELIMINAR TAREA ===\n");
-                let tareas=filtrarTodas(gestorTareas.getItems());
-                tareas.forEach(tarea => {
-                    console.log(tarea.toString());
-                    console.log("----------------------------------------");
-                });
-                console.log(`${tareas.length+1}. Salir del eliminar`)
+                    const idEliminar = pedirNumero("Ingrese la tarea que desea eliminar: ", 1 , tareas.length+1 , false)
 
-                const idEliminar = pedirNumero("Ingrese la tarea que desea eliminar: ", 1 , tareas.length+1 , false)
-
-                if(idEliminar == (tareas.length+1)){
-                    console.log("Volviendo al menu...");
+                    if(idEliminar == (tareas.length+1)){
+                        console.log("Volviendo al menu...");
+                        pausar();
+                    }
+                    const exito = gestorTareas.deleteItem(tareas[idEliminar-1].getId());
+                    if(exito){
+                        console.log("\n [OK] Tarea eliminada exitosamente.");
+                    }  else{
+                        console.log("\n [!] Error al eliminar la tarea.");
+                    }
                     pausar();
                 }
-                const exito = gestorTareas.deleteItem(tareas[idEliminar-1].getId());
-                if(exito){
-                    console.log("\n [OK] Tarea eliminada exitosamente.");
-                }  else{
-                    console.log("\n [!] Error al eliminar la tarea.");
-                }
-                
-                pausar();
-
             break;
             case 5:
-                input=pedirNumero(menuEstad(), 1 , 4 , false);
-                manejoEstadisticas(input, gestorTareas);
-                pausar();
+                if(gestorTareas.getItems().length == 0){
+                    console.log(" No hay tareas disponibles.");
+                    pausar();
+                }else{
+                    input=pedirNumero(menuEstad(), 1 , 4 , false);
+                    manejoEstadisticas(input, gestorTareas);
+                    pausar();
+                }
+                
             break;
             case 6:
-                input=pedirNumero(menuConsultas(), 1 , 3 , false);
-
+                if(gestorTareas.getItems().length == 0){
+                    console.log(" No hay tareas disponibles.");
+                    pausar();
+                }else{
+                    input=pedirNumero(menuConsultas(), 1 , 5 , false);
+                    manejoConsultas(input, gestorTareas);
+                    pausar();
+                }
             break;
             case 7:
                 console.log("\nCerrando aplicación... ¡Hasta luego!\n");
